+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Made `Table` component responsive by default.  This was the intended design
+  for this component.
+
 ## [0.1.2] - 2018-02-19
 
 ### Addded